style: Rename to applyEffects, because it is technically not middleware
